split the POST cognitive how-to card out of PostCognitiveDrillRunner (#4755)
PostDrillConfig and PostSessionLauncher were importing the 1900-line drill
runner purely to reach static help text, which inverted the dependency
direction for two surfaces that never run a drill.

The tutorial copy, the n-back worked example, the card, its standalone preview
and the button that opens it now live in CognitiveDrillTutorial.jsx. Only
DrillTutorialGate stays with the runner, since it gates the runner's own mount.
No back-compat re-exports: both consumers point at the new module directly.

Behavior is unchanged; the split is code organization only.
